backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
google-generativeai==0.3.2
google-genai==0.2.0
pydantic==2.5.3
tenacity==8.2.3
requests==2.31.0
gunicorn==21.2.0
